+++ EDDI/ChangeLog.md
@@ -6,7 +6,14 @@ Full details of the variables available for each noted event, and VoiceAttack in
   * Core
     * Updated system definition to include new variable `lastVisitSeconds`. 
   * Speech responder
+    * Updated the following events to include new property `stationtype`:
+      * `Docking cancelled`
+      * `Docking requested`
+      * `Docking granted`
+      * `Docking requested`
+      * `Docking timed out`
     * Updated `Commodity sales check` script to make use of `CommodityMarketDetails()` function.
+    * Updated `Docking granted` script to make use of of `stationtype` property.
     * Fixed a bug that had made `{ship.role}` inaccessible via scripts.
     * Updated `FSD engaged` script to correct a bug that was preventing sub-function `System report` from ever running.
   * Added the following Cottle function, documented in [the SpeechResponder documentation](https://github.com/EDCD/EDDI/blob/master/SpeechResponder/Help.md):